PAPP-7032 add 3 query actions
Allow flexible querying using the whole Censys query language by
support "query ip", "query domain", and "query certificate". Each
action allows the user to query one of the three main datasets in
Census, and get back a set of records, each of which can be passed to
the "lookup *" actions for more information.